voucher self signed and refactoring, generating bulk signals with promise.all
