+++ lib/voucher.env.ts
@@ -22,7 +22,7 @@ export type VoucherEnv = z.infer<typeof Env>;
 export const VoucherTypologies = z.enum(["PRODUCER", "CONSUMER"]);
 export type VoucherTypologies = z.infer<typeof VoucherTypologies>;
 
-export function getVoucherDataBy(voucherType: VoucherTypologies): VoucherEnv {
+export function getVocherEnvBy(voucherType: VoucherTypologies): VoucherEnv {
   const voucherConfigData = {};
   const path = `.env.${nodeEnv}.voucher.${voucherType.toLowerCase()}`;
   dotenv.config({
